feat(api): move typemap.ts, context.ts to folder called types
