+++ example/src/resolvers/Amenities.ts
@@ -1,5 +1,5 @@
 import { IAmenities } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 
 export interface AmenitiesParent {
   airConditioning: boolean;
