+++ example/src/resolvers/AuthPayload.ts
@@ -1,5 +1,5 @@
 import { IAuthPayload } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 import { UserParent } from "./User";
 
 export interface AuthPayloadParent {
