+++ example/src/resolvers/Booking.ts
@@ -1,5 +1,5 @@
 import { IBooking } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 import { UserParent } from "./User";
 import { PlaceParent } from "./Place";
 import { PaymentParent } from "./Payment";
