+++ example/src/resolvers/City.ts
@@ -1,5 +1,5 @@
 import { ICity } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 
 export interface CityParent {
   id: string;
