+++ example/src/resolvers/CreditCardInformation.ts
@@ -1,5 +1,5 @@
 import { ICreditCardInformation } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 import { PaymentAccountParent } from "./PaymentAccount";
 
 export interface CreditCardInformationParent {
