+++ example/src/resolvers/ExperienceCategory.ts
@@ -1,5 +1,5 @@
 import { IExperienceCategory } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 import { ExperienceParent } from "./Experience";
 
 export interface ExperienceCategoryParent {
