+++ example/src/resolvers/ExperiencesByCity.ts
@@ -1,5 +1,5 @@
 import { IExperiencesByCity } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 import { ExperienceParent } from "./Experience";
 import { CityParent } from "./City";
