+++ example/src/resolvers/GuestRequirements.ts
@@ -1,5 +1,5 @@
 import { IGuestRequirements } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 
 export interface GuestRequirementsParent {
   govIssuedId: boolean;
