+++ example/src/resolvers/Home.ts
@@ -1,5 +1,5 @@
 import { IHome } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 import { PictureParent } from "./Picture";
 
 export interface HomeParent {
