+++ example/src/resolvers/HouseRules.ts
@@ -1,5 +1,5 @@
 import { IHouseRules } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 
 export interface HouseRulesParent {
   additionalRules?: string;
