+++ example/src/resolvers/Mutation.ts
@@ -1,5 +1,5 @@
 import { IMutation } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 
 export interface MutationParent {}
