+++ example/src/resolvers/MutationResult.ts
@@ -1,5 +1,5 @@
 import { IMutationResult } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 
 export interface MutationResultParent {
   success: boolean;
