+++ example/src/resolvers/Payment.ts
@@ -1,5 +1,5 @@
 import { IPayment } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 import { BookingParent } from "./Booking";
 import { PaymentAccountParent } from "./PaymentAccount";
