+++ example/src/resolvers/PlaceViews.ts
@@ -1,5 +1,5 @@
 import { IPlaceViews } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 
 export interface PlaceViewsParent {
   id: string;
