+++ example/src/resolvers/Policies.ts
@@ -1,5 +1,5 @@
 import { IPolicies } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 
 export interface PoliciesParent {
   checkInEndTime: number;
