+++ example/src/resolvers/Query.ts
@@ -1,5 +1,5 @@
 import { IQuery } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 
 export interface QueryParent {}
