+++ example/src/resolvers/Reservation.ts
@@ -1,5 +1,5 @@
 import { IReservation } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 import { PictureParent } from "./Picture";
 import { LocationParent } from "./Location";
