+++ example/src/resolvers/Review.ts
@@ -1,5 +1,5 @@
 import { IReview } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 
 export interface ReviewParent {
   accuracy: number;
