+++ example/src/resolvers/Viewer.ts
@@ -1,5 +1,5 @@
 import { IViewer } from "../generated/resolvers";
-import { Types } from "./typemap";
+import { Types } from "./types/typemap";
 import { UserParent } from "./User";
 import { BookingParent } from "./Booking";
